// src/data/technology/technology-data.enum.ts
export enum TechnologyGroupDataEnum {
  frontend = 'frontend',
  programmingLanguage = 'programmingLanguage',
  framework = 'framework',
  library = 'library',
  database = 'database',
  test = 'test',
  tool = 'tool',
  operatingSystem = 'operatingSystem',
  architecture = 'architecture',
  ipaas = 'ipaas',
  rpa = 'rpa',
  documentGeneration = 'documentGeneration',
}

export enum TechnologyKindDataEnum {
  // Frontend
  html = 'html',
  css = 'css',
  sass = 'sass',
  storybook = 'storybook',
  // Programming Language
  javascript = 'javascript',
  typescript = 'typescript',
  // Framework
  angular = 'angular',
  // Library
  ngrx = 'ngrx',
  rxjs = 'rxjs',
  primeng = 'primeng',
  // Database
  sql = 'sql',
  microsoftSqlServer = 'microsoftSqlServer',
  // Test
  jest = 'jest',
  postman = 'postman',
  soapui = 'soapui',
  // Tool
  nodejs = 'nodejs',
  nx = 'nx',
  docker = 'docker',
  git = 'git',
  github = 'github',
  gitlab = 'gitlab',
  npm = 'npm',
  pnpm = 'pnpm',
  eslint = 'eslint',
  prettier = 'prettier',
  stylelint = 'stylelint',
  webstorm = 'webstorm',
  visualStudioCode = 'visualStudioCode',
  // Operating System
  windows = 'windows',
  linux = 'linux',
  // Architecture
  enterpriseArchitect = 'enterpriseArchitect',
  // iPaaS
  snapLogic = 'snapLogic',
  // RPA
  uiPath = 'uiPath',
  // Document Generation
  goSmartDesigner = 'goSmartDesigner',
}